refactor: split Home component into smaller files
