+++ rs-react-app/components/CardList.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import Card from './Card';
 import { Character } from '@/types';
