+++ rs-react-app/next.config.ts
@@ -2,7 +2,13 @@ import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
 	images: {
-		domains: ['rickandmortyapi.com'],
+		remotePatterns: [
+			{
+				protocol: 'https',
+				hostname: 'rickandmortyapi.com',
+				pathname: '/api/character/avatar/**',
+			},
+		],
 	},
 };
